+++ templates/ui-package/src/app/center-header/center-header.component.ts
@@ -1,0 +1,12 @@
+import {Component, OnInit} from '@angular/core';
+
+@Component({
+  selector: 'app-center-header',
+  templateUrl: './center-header.component.html',
+  styleUrls: ['./center-header.component.scss']
+})
+export class CenterHeaderComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit() {}
+}