+++ templates/ui-package/src/index.d.ts
@@ -1,0 +1,7 @@
+// so the typescript compiler doesn't complain about the global config object
+// app configuration injected from shell-ui
+declare var APP_CONF: any;
+// project build version injected from shell-ui
+declare var APP_BUILD_VERSION: string;
+// project app type injected from shell-ui
+declare var APP_TYPE: any;